refactor: replace old reader impl with new
