+++ test/reader.test.js
@@ -1,20 +1,13 @@
 'use strict';
 
-const bundleJS = require('../');
+const bundleJS = require('../lib/reader');
 const path = require('path');
 const Sink = require('@asset-pipe/sink-fs');
 const prettier = require('prettier');
+const { join } = require('path');
+const { remove } = require('fs-extra');
+const FOLDER = join(__dirname, 'test-assets-reader');
 const vm = require('vm');
-
-const SIMPLE_FEED_A_HASH =
-    'c26fae70ab34dd08230cf2ec30dbc4ed70d1f6616f38f7b3101e6618c9d43ebf';
-const SIMPLE_FEED_B_HASH =
-    '5b54fc816e7c9157d5e6948b6a60cca9ddbb92b81e8d41fd085f74a15645c703';
-const SIMPLE_FEED_C_HASH =
-    '5128670188c99f2817fdef2293c0157d430e6b90324ee95a55dbb676e7d64d68';
-
-const FEED_A_HASH = '9fab6ddd2a66a14b8b2ab6373fefa8987fabafb1';
-const FEED_B_HASH = '9fab6ddd2a66a14b8b2ab6373fefa8987fabafb2';
 
 function getExecutionOrder(bundle) {
     const lines = bundle.split('\n').filter(Boolean);
@@ -28,161 +21,214 @@ function getExecutionOrder(bundle) {
     return JSON.parse(order);
 }
 
-test.skip('should concat 3 files', async () => {
+beforeEach(() => remove(FOLDER));
+afterAll(() => remove(FOLDER));
+
+test('should successfully bundle 2 feeds', async () => {
+    const sink = new Sink({ path: path.join(__dirname, 'mock') });
+
+    const feedA = JSON.parse(await sink.get('feed.c.json'));
+    const feedB = JSON.parse(await sink.get('feed.d.json'));
+
+    const content = await bundleJS([feedA, feedB], {
+        directory: FOLDER,
+    });
+
+    expect(prettier.format(content)).toMatchSnapshot();
+});
+
+test('should dedupe', async () => {
+    const sink = new Sink({ path: path.join(__dirname, 'mock') });
+
+    const feedA = JSON.parse(await sink.get('feed.e.json'));
+    const feedB = JSON.parse(await sink.get('feed.f.json'));
+
+    const content = await bundleJS([feedA, feedB], {
+        directory: FOLDER,
+    });
+
+    expect(prettier.format(content)).toMatchSnapshot();
+});
+
+test('should handle node_modules dependencies', async () => {
     const sink = new Sink({ path: path.join(__dirname, 'mock') });
 
     const feedA = JSON.parse(await sink.get('simple.a.json'));
     const feedB = JSON.parse(await sink.get('simple.b.json'));
     const feedC = JSON.parse(await sink.get('simple.c.json'));
 
-    const content = await bundleJS([feedA, feedB, feedC]);
-    const executionOrder = getExecutionOrder(content);
+    const content = await bundleJS([feedA, feedB, feedC], {
+        directory: FOLDER,
+    });
 
-    expect(executionOrder).toHaveLength(3);
-    expect(executionOrder).toEqual([
-        SIMPLE_FEED_A_HASH,
-        SIMPLE_FEED_B_HASH,
-        SIMPLE_FEED_C_HASH,
-    ]);
     expect(prettier.format(content)).toMatchSnapshot();
 });
 
-test.skip('should concat 2 files', async () => {
+test('should concat 3 files', async () => {
+    const sink = new Sink({ path: path.join(__dirname, 'mock') });
+
+    const feedA = JSON.parse(await sink.get('simple.a.json'));
+    const feedB = JSON.parse(await sink.get('simple.b.json'));
+    const feedC = JSON.parse(await sink.get('simple.c.json'));
+
+    const content = await bundleJS([feedA, feedB, feedC], {
+        directory: FOLDER,
+    });
+    const executionOrder = getExecutionOrder(content);
+
+    expect(executionOrder).toHaveLength(3);
+    expect(executionOrder).toEqual([10, 4, 19]);
+    expect(prettier.format(content)).toMatchSnapshot();
+});
+
+test('should concat 2 files', async () => {
     const sink = new Sink({ path: path.join(__dirname, 'mock') });
 
     const feedA = JSON.parse(await sink.get('feed.a.json'));
     const feedB = JSON.parse(await sink.get('feed.b.json'));
 
-    const content = await bundleJS([feedA, feedB]);
+    const content = await bundleJS([feedA, feedB], {
+        directory: FOLDER,
+    });
 
     const executionOrder = getExecutionOrder(content);
 
     expect(executionOrder).toHaveLength(2);
-    expect(executionOrder).toEqual([FEED_A_HASH, FEED_B_HASH]);
+    expect(executionOrder).toEqual([3, 9]);
     expect(prettier.format(content)).toMatchSnapshot();
 });
 
-test.skip('should concat 1 file', async () => {
+test('should concat 1 file', async () => {
     const sink = new Sink({ path: path.join(__dirname, 'mock') });
 
     const feedA = JSON.parse(await sink.get('feed.a.json'));
 
-    const content = await bundleJS([feedA]);
+    const content = await bundleJS([feedA], {
+        directory: FOLDER,
+    });
 
     const executionOrder = getExecutionOrder(content);
 
     expect(executionOrder).toHaveLength(1);
-    expect(executionOrder).toEqual([FEED_A_HASH]);
+    expect(executionOrder).toEqual([3]);
     expect(prettier.format(content)).toMatchSnapshot();
 });
 
-test.skip('code reach 1 entry point', async () => {
-    const result = await bundleJS([
+test('code reach 1 entry point', async () => {
+    const result = await bundleJS(
         [
-            {
-                entry: true,
-                id: 'a',
-                source: 'spy(1234);',
-            },
+            [
+                {
+                    entry: true,
+                    id: 'a',
+                    file: '',
+                    source: 'spy(1234);',
+                    deps: {},
+                },
+            ],
         ],
-    ]);
+        {
+            directory: FOLDER,
+        }
+    );
     const spy = jest.fn();
     vm.runInNewContext(result, { spy });
     expect(spy).toHaveBeenCalledTimes(1);
-    expect(spy.mock.calls).toMatchSnapshot();
-});
-
-test.skip('code reach 3 entry point', async () => {
-    const result = await bundleJS([
-        [
-            {
-                entry: true,
-                id: 'a',
-                source: 'spy("a");',
-            },
-        ],
-        [
-            {
-                entry: true,
-                id: 'b',
-                source: 'spy("b");',
-            },
-        ],
-        [
-            {
-                entry: true,
-                id: 'c',
-                source: 'spy("c");',
-            },
-        ],
-    ]);
-    const spy = jest.fn();
-    vm.runInNewContext(result, { spy });
-    expect(spy).toHaveBeenCalledTimes(3);
-    expect(spy.mock.calls).toMatchSnapshot();
-});
-
-test.skip('should exclude/dedupe common modules', async () => {
-    const result = await bundleJS([
-        [
-            {
-                entry: true,
-                id: 'a',
-                source: 'require("./c");spy("a");',
-                deps: { './c': 'c' },
-            },
-            {
-                id: 'c',
-                source: 'spy("c");',
-                deps: {},
-            },
-        ],
-        [
-            {
-                entry: true,
-                id: 'b',
-                source: 'require("./c");spy("b");',
-                deps: { './c': 'c' },
-            },
-            {
-                id: 'c',
-                source: 'spy("c");',
-                deps: {},
-            },
-        ],
-    ]);
-    const spy = jest.fn();
-    vm.runInNewContext(result, { spy });
-
     expect(spy).toMatchSnapshot();
-    expect(prettier.format(result)).toMatchSnapshot();
 });
 
-test.skip('should error if no feed content', async () => {
+test('should error if no feed content', async () => {
     expect(bundleJS()).rejects.toMatchSnapshot();
     expect(bundleJS([])).rejects.toMatchSnapshot();
     expect(bundleJS([[]])).rejects.toMatchSnapshot();
-    expect(bundleJS([[{}]])).rejects.toMatchSnapshot();
 });
 
-test.skip('when deps dont match, should not dedupe', async () => {
-    const SinkFs = require('@asset-pipe/sink-fs');
-    const sink = new SinkFs({ path: `${__dirname}/mock` });
-    const feedC = JSON.parse(await sink.get('feed.c.json'));
-    const feedD = JSON.parse(await sink.get('feed.d.json'));
+test('code reaches 3 entry points', async () => {
+    const result = await bundleJS(
+        [
+            [
+                {
+                    entry: true,
+                    id: 'a',
+                    source: 'spy("a");',
+                    file: '',
+                    deps: {},
+                },
+            ],
+            [
+                {
+                    entry: true,
+                    id: 'b',
+                    source: 'spy("b");',
+                    file: '',
+                    deps: {},
+                },
+            ],
+            [
+                {
+                    entry: true,
+                    id: 'c',
+                    source: 'spy("c");',
+                    file: '',
+                    deps: {},
+                },
+            ],
+        ],
+        {
+            directory: FOLDER,
+        }
+    );
+    const spy = jest.fn();
+    vm.runInNewContext(result, { spy });
 
-    const bundle = await bundleJS([feedC, feedD]);
-    // console.log(bundle);
-    expect(bundle.match(/my module/g)).toHaveLength(2);
+    expect(prettier.format(result)).toMatchSnapshot();
+    expect(spy).toHaveBeenCalledTimes(3);
+    expect(spy).toMatchSnapshot();
 });
 
-test.skip('when deps do match, should dedupe', async () => {
-    const SinkFs = require('@asset-pipe/sink-fs');
-    const sink = new SinkFs({ path: `${__dirname}/mock` });
-    const feedE = JSON.parse(await sink.get('feed.e.json'));
-    const feedF = JSON.parse(await sink.get('feed.f.json'));
+test('bundling dedupes common modules', async () => {
+    const result = await bundleJS(
+        [
+            [
+                {
+                    entry: true,
+                    id: 'a',
+                    source: 'require("./c"); spy("a");',
+                    deps: { './c': 'c' },
+                    file: './a.js',
+                },
+                {
+                    id: 'c',
+                    source: 'spy("c");',
+                    deps: {},
+                    file: './c.js',
+                },
+            ],
+            [
+                {
+                    entry: true,
+                    id: 'b',
+                    source: 'require("./c"); spy("b");',
+                    deps: { './c': 'c' },
+                    file: './b.js',
+                },
+                {
+                    id: 'c',
+                    source: 'spy("c");',
+                    deps: {},
+                    file: './c.js',
+                },
+            ],
+        ],
+        {
+            directory: FOLDER,
+        }
+    );
+    const spy = jest.fn();
+    vm.runInNewContext(result, { spy });
 
-    const bundle = await bundleJS([feedE, feedF]);
-
-    expect(bundle.match(/my module/g)).toHaveLength(1);
+    expect(prettier.format(result)).toMatchSnapshot();
+    expect(spy).toMatchSnapshot();
+    expect(spy).toHaveBeenCalledTimes(4);
+    expect(prettier.format(result)).toMatchSnapshot();
 });
